fix widget rebuilding error
The following assertion was thrown while dispatching notifications for LoggingService:
setState() or markNeedsBuild() called during build.
This _InheritedProviderScope<LoggingService?> widget cannot be marked as needing to build because
the framework is already in the process of building widgets. A widget can be marked as needing to be
built during the build phase only if one of its ancestors is currently building. This exception is
allowed because the framework builds parent widgets before children, which means a dirty descendant
will always be built. Otherwise, the framework might not visit this widget during this build phase.
The widget on which setState() or markNeedsBuild() was called was:
  _InheritedProviderScope<LoggingService?>
The widget which was currently being built when the offending call was made was:
  Builder